Rename evaluator
Closes #684
Trigger CI
